add lottery and queue logic

+++ src/main/java/application/LotteryService.java
@@ -30,7 +30,7 @@ public class LotteryService {
         this.scheduler = Executors.newScheduledThreadPool(10);
     }
 
-    public Response<Boolean> createLottery(String token, String eventId, int capacity, LocalDateTime registerWindow, double expirationTime) {
+    public Response<Boolean> createLottery(String token, String eventId, int capacity, LocalDateTime registerWindow, long expirationTime) {
         logger.log(Level.INFO, "createLottery called");
 
         // check valid token
